+++ packages/datagraphic/stories/scales.stories.js
@@ -1,4 +1,5 @@
 import ScalePointStory from "./scales/ScalePoint.svelte";
+import LogStory from "./scales/Log.svelte";
 
 export default {
   title: "Scales",
@@ -6,4 +7,8 @@ export default {
 
 export const ScalePoint = () => ({
   Component: ScalePointStory,
+});
+
+export const Log = () => ({
+  Component: LogStory,
 });